Reuse CanonicalConfig between controllers
