+++ operators/pkg/controller/common/settings/canonical_config_test.go
