+++ operators/pkg/controller/elasticsearch/mutation/comparison/config.go
@@ -7,10 +7,11 @@ package comparison
 import (
 	"fmt"
 
+	commonsettings "github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/settings"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/settings"
 )
 
-func compareConfigs(actual *settings.CanonicalConfig, expected *settings.CanonicalConfig) Comparison {
+func compareConfigs(actual *commonsettings.CanonicalConfig, expected *commonsettings.CanonicalConfig) Comparison {
 	// check for settings in actual that do not match expected
 	diff := actual.Diff(expected, toIgnore)
 	if len(diff) == 0 {
